Create Reverse Shuffle Merge
see my easy c++ code after reading the question and seeing the test case its simple we have to iterate from last and have to take smallest string. to do that follow this. 1.simply take total frequency of character and how much we need which is half of frequency of every character. 2. take a stack of char and insert if its empty or check if this character we need then before inserting just remove all charcater bigger than this which will repeat again. 3. now in the stack we have our string.